move enumeration tests to own file
